transaction-extractor/requirements.txt
# Core dependencies
pydantic==2.5.0
python-dotenv==1.0.0

# File processing
pdfplumber==0.10.3
PyPDF2==3.0.1
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# AI API (Groq uses OpenAI-compatible client)
openai>=1.0.0

# Utilities
python-dateutil==2.8.2

